+++ source/blender/editors/include/UI_interface_icons.h
@@ -1,5 +1,5 @@
 /*
- * $Id: UI_interface_icons.h 37210 2011-06-05 14:00:06Z campbellbarton $
+ * $Id: UI_interface_icons.h 40539 2011-09-25 12:33:51Z ender79 $
  *
  * ***** BEGIN GPL LICENSE BLOCK *****
  *
@@ -50,8 +50,8 @@ typedef struct IconFile {
 #define ICON_DEFAULT_HEIGHT 16
 #define ICON_DEFAULT_WIDTH	16
 
-#define ICON_DEFAULT_HEIGHT_SCALE	(UI_UNIT_Y * 0.8f)
-#define ICON_DEFAULT_WIDTH_SCALE	(UI_UNIT_X * 0.8f)
+#define ICON_DEFAULT_HEIGHT_SCALE ((int)(UI_UNIT_Y * 0.8f))
+#define ICON_DEFAULT_WIDTH_SCALE  ((int)(UI_UNIT_X * 0.8f))
 
 #define PREVIEW_DEFAULT_HEIGHT 96
